+++ src/pages/Lend/LoanChannelCategoryClimateExperiment.vue
@@ -62,12 +62,11 @@
 			<div v-for="(channel, index) in secondaryChannels" :key="index" class="tw-mt-6">
 				<kiva-classic-single-category-carousel
 					:id="`carousel-${channel.name}` | changeCase('paramCase')"
+					:prefetched-selected-channel="channel"
 					:climate-challenge="isEcoChallengeExpShown"
-					:loan-channel-id="channel.id"
-					:loan-channel-name="channel.name"
-					:loan-channel-description="channel.description"
 					:loan-display-settings="loanDisplaySettings"
 					:lend-now-button="true"
+					:query-context="ecoExpQueryContext"
 				/>
 			</div>
 		</kv-page-container>
@@ -82,6 +81,7 @@ import lendFilterExpMixin from '@/plugins/lend-filter-page-exp-mixin';
 import loanChannelQueryMapMixin from '@/plugins/loan-channel-query-map';
 import ViewToggle from '@/components/LoansByCategory/ViewToggle';
 import KivaClassicSingleCategoryCarousel from '@/components/LoanCollections/KivaClassicSingleCategoryCarousel';
+import { FLSS_ORIGIN_CATEGORY } from '@/util/flssUtils';
 import {
 	preFetchChannel,
 	getCachedChannel
@@ -94,6 +94,7 @@ import {
 	mdiClose
 } from '@mdi/js';
 import IconCalendar from '@/assets/icons/inline/eco-challenge/calendar.svg';
+import gql from 'graphql-tag';
 import KvPageContainer from '~/@kiva/kv-components/vue/KvPageContainer';
 import KvTextLink from '~/@kiva/kv-components/vue/KvTextLink';
 import KvMaterialIcon from '~/@kiva/kv-components/vue/KvMaterialIcon';
@@ -113,6 +114,18 @@ const secondaryEcoLoanChannels = loanChannelQueryMapMixin.data().loanChannelQuer
 
 const secondaryEcoLoanChannelIds = secondaryEcoLoanChannels.map(channel => channel.id);
 
+const mainCategoryQuery = gql`
+	query mainCategoryQuery($ids: [Int]) {
+		lend {
+			loanChannelsById (ids: $ids) {
+				id
+				name
+				description
+				url
+			}
+		}
+	}
+`;
 export default {
 	name: 'LoanChannelCategoryClimateExperiment',
 	components: {
@@ -144,6 +157,7 @@ export default {
 			isEcoChallengeExpShown: false,
 			showHowItWorks: true,
 			mdiClose,
+			ecoExpQueryContext: FLSS_ORIGIN_CATEGORY
 		};
 	},
 	computed: {
@@ -193,14 +207,26 @@ export default {
 			// Get page limit and offset
 			const limit = 9;
 			const offset = 0;
-			return preFetchChannel(
-				client,
-				// Access map directly since SSR doesn't have mixins available
-				loanChannelQueryMapMixin.data().loanChannelQueryMap,
-				targetedLoanChannelURL,
-				// Build loanQueryVars since SSR doesn't have same context
-				{ ids: [...secondaryEcoLoanChannelIds, targetedLoanChannelID], limit, offset }
-			);
+
+			const promisesArrayOfSecondaryChannels = secondaryEcoLoanChannelIds.map(channelId => {
+				return preFetchChannel(
+					client,
+					loanChannelQueryMapMixin.data().loanChannelQueryMap,
+					loanChannelQueryMapMixin.data().loanChannelQueryMap.find(channel => channel.id === channelId)?.url,
+					{
+						ids: [channelId], limit, offset, origin: FLSS_ORIGIN_CATEGORY
+					}
+				);
+			});
+
+			// Just get the description and name of main loan channel
+			const mainChannel = client.query({
+				query: mainCategoryQuery,
+				variables: {
+					ids: [targetedLoanChannelID],
+				},
+			});
+			return Promise.all([mainChannel, ...promisesArrayOfSecondaryChannels]);
 		}
 	},
 	created() {
@@ -209,21 +235,36 @@ export default {
 			.find(channel => channel.url === this.targetedLoanChannelURL)?.id;
 
 		// Prevent pop-in by loading data from the Apollo cache manually here instead of just using the subscription
-		const baseData = getCachedChannel(
-			this.apollo,
-			this.loanChannelQueryMap,
-			this.targetedLoanChannelURL,
-			{
-				ids: [...secondaryEcoLoanChannelIds, this.targetedLoanChannelID],
-				limit: 9,
-				offset: 0,
-				basketId: this.cookieStore.get('kvbskt'),
-			}
-		);
-		this.loanChannel = baseData?.lend?.loanChannelsById.find(channel => channel.id === this.targetedLoanChannelID);
+		const mainChannelData = this.apollo.readQuery({
+			query: mainCategoryQuery,
+			variables: {
+				ids: [this.targetedLoanChannelID],
+			},
+		});
+		this.loanChannel = mainChannelData?.lend?.loanChannelsById
+			.find(channel => channel.id === this.targetedLoanChannelID);
 
-		this.secondaryEcoLoanChannelsResponse = baseData?.lend?.loanChannelsById
-			.filter(channel => channel.id !== this.targetedLoanChannelID) ?? [];
+		// Get secondary channels
+		// Prevent pop-in by loading data from the Apollo cache manually here instead of just using the subscription
+		const secondaryChannelResponse = secondaryEcoLoanChannelIds.map(channelId => {
+			return getCachedChannel(
+				this.apollo,
+				this.loanChannelQueryMap,
+				this.loanChannelQueryMap.find(channel => channel.id === channelId)?.url,
+				{
+					ids: [channelId],
+					limit: 9,
+					offset: 0,
+					origin: FLSS_ORIGIN_CATEGORY,
+					basketId: this.cookieStore.get('kvbskt'),
+				}
+			);
+		});
+
+		// Use reduce to combine lend.loanChannelById property into 1 array
+		this.secondaryEcoLoanChannelsResponse = secondaryChannelResponse.reduce((acc, channel) => {
+			return [...acc, ...channel.lend.loanChannelsById];
+		}, []);
 
 		// filter out any secondary channels that do not have loans
 		this.secondaryEcoLoanChannelsResponse = this.secondaryEcoLoanChannelsResponse