+++ src/pages/Lend/LoanChannelCategoryControl.vue
@@ -2,7 +2,15 @@
 	<div class="tw-relative">
 		<div class="row">
 			<div class="small-12 columns heading-region">
-				<view-toggle browse-url="/lend-by-category" :filter-url="filterUrl" />
+				<view-toggle v-if="!enableQuickFilters" browse-url="/lend-by-category" :filter-url="filterUrl" />
+				<router-link
+					v-else
+					:to="filterUrl"
+					class="tw-text-action tw-flex tw-items-center tw-float-right"
+				>
+					<img class="tw-w-2 tw-mr-1" src="@/assets/images/tune.svg">
+					Advanced filters
+				</router-link>
 				<p class="tw-text-small">
 					<router-link to="/lend-by-category">
 						All Loans
@@ -120,6 +128,7 @@ import {
 	formatSortOptions,
 	transformIsoCodes,
 } from '@/util/loanSearch/filterUtils';
+import { FLSS_ORIGIN_CATEGORY } from '@/util/flssUtils';
 import QuickFilters from '@/components/LoansByCategory/QuickFilters/QuickFilters';
 
 const defaultLoansPerPage = 12;
@@ -240,7 +249,7 @@ export default {
 			return _get(this.loanChannel, 'description') || null;
 		},
 		loans() {
-			return _get(this.loanChannel, 'loans.values') || [];
+			return (this.loanChannel?.loans?.values ?? []).filter(loan => loan !== null);
 		},
 		firstLoan() {
 			// Handle an edge case where a backend error could lead to a null loan
@@ -261,6 +270,7 @@ export default {
 				limit: this.limit,
 				offset: this.offset,
 				basketId: this.cookieStore.get('kvbskt'),
+				origin: FLSS_ORIGIN_CATEGORY
 			};
 		},
 		filterUrl() {
@@ -320,7 +330,12 @@ export default {
 					loanChannelQueryMapMixin.data().loanChannelQueryMap,
 					targetedLoanChannelURL,
 					// Build loanQueryVars since SSR doesn't have same context
-					{ ids: [targetedLoanChannelID], limit, offset },
+					{
+						ids: [targetedLoanChannelID],
+						limit,
+						offset,
+						origin: FLSS_ORIGIN_CATEGORY
+					},
 					this.selectedQuickFilters
 				);
 			});
@@ -550,7 +565,7 @@ export default {
 		},
 		async fetchFacets(loanSearchState = {}) {
 			// TODO: Prevent this from running on every query (not needed for sorting and paging)
-			const { isoCodes } = await runFacetsQueries(this.apollo, loanSearchState);
+			const { isoCodes } = await runFacetsQueries(this.apollo, loanSearchState, FLSS_ORIGIN_CATEGORY);
 
 			// Merge all facet options with filtered options
 			const facets = {