+++ CHANGELOG.md
@@ -1,3 +1,80 @@
+## [2.327.0](https://github.com/kiva/ui/compare/v2.326.0...v2.327.0) (2022-10-11)
+
+
+### Features
+
+* new badge for quick filter experiment was added ([18effb8](https://github.com/kiva/ui/commit/18effb80ccbdf769fdc7e0570eec1e2e8557de59))
+
+## [2.326.0](https://github.com/kiva/ui/compare/v2.325.0...v2.326.0) (2022-10-10)
+
+
+### Features
+
+* add origin query param to lbc eco refactor ([1a28f6e](https://github.com/kiva/ui/commit/1a28f6e6d57d7b11314f3ae1f3619eab7915df5e))
+
+## [2.325.0](https://github.com/kiva/ui/compare/v2.324.1...v2.325.0) (2022-10-10)
+
+
+### Features
+
+* copy change on eco thanks page ([abc2415](https://github.com/kiva/ui/commit/abc2415f14495fff03952bab687c3945052eaaff))
+
+### [2.324.1](https://github.com/kiva/ui/compare/v2.324.0...v2.324.1) (2022-10-10)
+
+
+### Bug Fixes
+
+* add flss origin constants and use in all flss queries ([060d7a3](https://github.com/kiva/ui/commit/060d7a3e3a688cb57391b5da3a256c2a76851733))
+* pass query origin context on all flss queries ([21c89ba](https://github.com/kiva/ui/commit/21c89ba5a3eed4fe0fe9953d0820699b229627c5))
+
+## [2.324.0](https://github.com/kiva/ui/compare/v2.323.0...v2.324.0) (2022-10-07)
+
+
+### Features
+
+* add event tracking for eco challenge confirmation modal ([d0c872f](https://github.com/kiva/ui/commit/d0c872fce3981dd8689b9b40f0185cd718c28f84))
+
+## [2.323.0](https://github.com/kiva/ui/compare/v2.322.0...v2.323.0) (2022-10-07)
+
+
+### Features
+
+* extend FLSS exp now pulled ([beba440](https://github.com/kiva/ui/commit/beba4401e9094800e8cf500688a768041161de1d))
+
+## [2.322.0](https://github.com/kiva/ui/compare/v2.321.0...v2.322.0) (2022-10-07)
+
+
+### Features
+
+* tag loan search filter ([02861f5](https://github.com/kiva/ui/commit/02861f541628c2d9f697d20e33a2711a2c1e4356))
+* updated loan map for tagId ([8a571d1](https://github.com/kiva/ui/commit/8a571d10064f46be9b9aea61406b9e1a587c1687))
+
+
+### Bug Fixes
+
+* updated map prop ([40367fc](https://github.com/kiva/ui/commit/40367fcff414ad55f7bf848f1019ea0a04d86c2d))
+
+## [2.321.0](https://github.com/kiva/ui/compare/v2.320.1...v2.321.0) (2022-10-06)
+
+
+### Features
+
+* new advanced filters entrypoint was added for quick filters experiment ([a112703](https://github.com/kiva/ui/commit/a112703e73d52fed342f871908743b4b40b9eb67))
+
+### [2.320.1](https://github.com/kiva/ui/compare/v2.320.0...v2.320.1) (2022-10-06)
+
+
+### Bug Fixes
+
+* fix amend to previous commit ([3213d36](https://github.com/kiva/ui/commit/3213d3600deb9ed7d967b7ec91f1a9cc42b46664))
+
+## [2.320.0](https://github.com/kiva/ui/compare/v2.319.0...v2.320.0) (2022-10-06)
+
+
+### Features
+
+* remove querying for category info again for single category carousel ([1f9f0bc](https://github.com/kiva/ui/commit/1f9f0bc33a07378a203d5fb9122ba7f290b77c6e))
+
 ## [2.319.0](https://github.com/kiva/ui/compare/v2.318.1...v2.319.0) (2022-10-06)